refactor(F204): move url-safety to utils — generic SSRF protection, not weixin-specific
url-safety.ts provides DNS pinning, IPv4/IPv6 private range detection,
and safe external URL fetching. Any plugin that downloads external
resources needs this, so it belongs in shared utils, not under
domains/weixin-mp/.

[宪宪/claude-opus-4-6🐾]
